requirements.txt
sqlmodel
pymysql
cryptography
cassandra-driver